옷 걸기 기능 완료. 블루투스 통신 테스트 필요

+++ app/src/main/java/com/stac/hanghangtwo/util/Id.java
@@ -1,6 +1,6 @@
 package com.stac.hanghangtwo.util;
 
 public class Id {
-    public static byte id;
+    public static int id;
     public final static String uuid = "00001101-0000-1000-8000-00805f9b34fb";
 }
